Support more than redis for celery proc queue length.
Not all of kombu's backends have a _size() paramater. Notably, neither
of the AMQP backends do. In both cases (py-amqp and librabbitmq) we can
passively declare a queue (catching the does not exist error) and get
the number of queued messages from there.

Tested locally with Redis, py-amqp and librabbitmq.