Default to checking celery workers for active jobs
The process takes longer, but the behavior needs to be unbroken by
default, even at the cost of speed.

However, there are legitimate cases for attempting to speed up the
process, and those can be optimized by overriding the inspect_statuses
property on the proc.

Document how `inspect_statuses` works, and when it might be appropriate
to ignore a particular status.